norminett split

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: rucosta <rucosta@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/07 18:48:54 by rucosta           #+#    #+#             */
-/*   Updated: 2025/04/09 21:50:44 by rucosta          ###   ########.fr       */
+/*   Updated: 2025/04/29 21:29:08 by rucosta          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *c)
+#include "libft.h"
+
+size_t	ft_strlen(const char *c)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (c[len] != '\0')
